0. Refactoring sarcini si structura proiect
