--- src/main/resources/variantele.txt
@@ -1,1 +1,0 @@
-Rulati clasa InternshipUtils sa vedeti fiecare ce varianta aveti.
